Remove uproot_methods dependency
uproot_methods isn't in the official CMSSW release. So I was suggested
to remove it.

I tried to change the method to TTree.Draw(). It works fine standalone,
but showed conflict with TopTagger module with the TopTaggerInterface.

The old looping over event method in TTree is very slow .

Now I am still using the uproot to obtain the arrays and numpy histogram.
Just manually save to the TH1F and store. Hopefully it will work in general.